Move data-plane new trigger filters to subscriptionsapi package
