+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/subscriptionsapi/AllFilter.java
@@ -13,23 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi;
 
 import dev.knative.eventing.kafka.broker.dispatcher.Filter;
 import io.cloudevents.CloudEvent;
 
 import java.util.Set;
 
-public class NotFilter implements Filter {
+public class AllFilter implements Filter {
 
-  private Filter filter;
+  private final Set<Filter> filters;
 
-  public NotFilter(Filter filter) {
-    this.filter = filter;
+  public AllFilter(Set<Filter> filters) {
+    this.filters = filters;
   }
 
   @Override
   public boolean test(CloudEvent cloudEvent) {
-   return !filter.test(cloudEvent);
+    for (Filter filter : filters) {
+      if (!filter.test(cloudEvent)) {
+        return false;
+      }
+    }
+    return true;
   }
 }
