+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/main/ConsumerVerticleFactoryImpl.java
@@ -29,9 +29,9 @@ import dev.knative.eventing.kafka.broker.dispatcher.ResponseHandler;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.NoopResponseHandler;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.RecordDispatcherImpl;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.RecordDispatcherMutatorChain;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.ResourceContext;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.ResponseToHttpEndpointHandler;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.ResponseToKafkaTopicHandler;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.ResourceContext;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.BaseConsumerVerticle;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.CloudEventOverridesMutator;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.InvalidCloudEventInterceptor;
@@ -39,13 +39,13 @@ import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.KeyDeserialize
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.OffsetManager;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.OrderedConsumerVerticle;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.consumer.UnorderedConsumerVerticle;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.AllFilter;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.AnyFilter;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.AttributesFilter;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.NotFilter;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.PrefixFilter;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.SqlFilter;
-import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.SuffixFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.AllFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.AnyFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.NotFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.PrefixFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.SqlFilter;
+import dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi.SuffixFilter;
 import dev.knative.eventing.kafka.broker.dispatcher.impl.http.WebClientCloudEventSender;
 import io.cloudevents.CloudEvent;
 import io.micrometer.core.instrument.Counter;
@@ -155,59 +155,59 @@ public class ConsumerVerticleFactoryImpl implements ConsumerVerticleFactory {
 
     final BaseConsumerVerticle.Initializer initializer = (vertx, consumerVerticle) ->
       authProvider.getCredentials(resource).onSuccess(credentials -> {
-        KafkaClientsAuth.attachCredentials(consumerConfigs, credentials);
-        KafkaClientsAuth.attachCredentials(producerConfigs, credentials);
+          KafkaClientsAuth.attachCredentials(consumerConfigs, credentials);
+          KafkaClientsAuth.attachCredentials(producerConfigs, credentials);
 
-        final KafkaConsumer<Object, CloudEvent> consumer = createConsumer(vertx, consumerConfigs);
-        AutoCloseable metricsCloser = Metrics.register(consumer.unwrap());
+          final KafkaConsumer<Object, CloudEvent> consumer = createConsumer(vertx, consumerConfigs);
+          AutoCloseable metricsCloser = Metrics.register(consumer.unwrap());
 
-        final var egressConfig =
-          egress.hasEgressConfig() ?
-            egress.getEgressConfig() :
-            resource.getEgressConfig();
+          final var egressConfig =
+            egress.hasEgressConfig() ?
+              egress.getEgressConfig() :
+              resource.getEgressConfig();
 
-        final var egressSubscriberSender = createConsumerRecordSender(
-          vertx,
-          egress.getDestination(),
-          egressConfig
-        );
+          final var egressSubscriberSender = createConsumerRecordSender(
+            vertx,
+            egress.getDestination(),
+            egressConfig
+          );
 
-        final var egressDeadLetterSender = hasDeadLetterSink(egressConfig)
-          ? createConsumerRecordSender(vertx, egressConfig.getDeadLetter(), egressConfig)
-          : NO_DEAD_LETTER_SINK_SENDER;
+          final var egressDeadLetterSender = hasDeadLetterSink(egressConfig)
+            ? createConsumerRecordSender(vertx, egressConfig.getDeadLetter(), egressConfig)
+            : NO_DEAD_LETTER_SINK_SENDER;
 
-        final var filter = getFilter(egress);
+          final var filter = getFilter(egress);
 
-        final var responseHandler = getResponseHandler(egress,
-          () -> getResponseToKafkaTopicHandler(vertx, producerConfigs, resource),
-          () -> new ResponseToHttpEndpointHandler(createConsumerRecordSender(vertx, egress.getReplyUrl(), egressConfig)));
-        final var commitIntervalMs = Integer.parseInt(String.valueOf(consumerConfigs.get(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG)));
+          final var responseHandler = getResponseHandler(egress,
+            () -> getResponseToKafkaTopicHandler(vertx, producerConfigs, resource),
+            () -> new ResponseToHttpEndpointHandler(createConsumerRecordSender(vertx, egress.getReplyUrl(), egressConfig)));
+          final var commitIntervalMs = Integer.parseInt(String.valueOf(consumerConfigs.get(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG)));
 
-        final var recordDispatcher = new RecordDispatcherMutatorChain(
-          new RecordDispatcherImpl(
-            new ResourceContext(resource, egress),
-            filter,
-            egressSubscriberSender,
-            egressDeadLetterSender,
-            responseHandler,
-            new OffsetManager(vertx, consumer, eventsSentCounter::increment, commitIntervalMs),
-            ConsumerTracer.create(
-              ((VertxInternal) vertx).tracer(),
-              new KafkaClientOptions()
-                .setConfig(consumerConfigs)
-                // Make sure the policy is propagate for the manually instantiated consumer tracer
-                .setTracingPolicy(TracingPolicy.PROPAGATE)
+          final var recordDispatcher = new RecordDispatcherMutatorChain(
+            new RecordDispatcherImpl(
+              new ResourceContext(resource, egress),
+              filter,
+              egressSubscriberSender,
+              egressDeadLetterSender,
+              responseHandler,
+              new OffsetManager(vertx, consumer, eventsSentCounter::increment, commitIntervalMs),
+              ConsumerTracer.create(
+                ((VertxInternal) vertx).tracer(),
+                new KafkaClientOptions()
+                  .setConfig(consumerConfigs)
+                  // Make sure the policy is propagate for the manually instantiated consumer tracer
+                  .setTracingPolicy(TracingPolicy.PROPAGATE)
+              ),
+              Metrics.getRegistry()
             ),
-            Metrics.getRegistry()
-          ),
-          new CloudEventOverridesMutator(resource.getCloudEventOverrides())
-        );
+            new CloudEventOverridesMutator(resource.getCloudEventOverrides())
+          );
 
-        // Set all the built objects in the consumer verticle
-        consumerVerticle.setRecordDispatcher(recordDispatcher);
-        consumerVerticle.setConsumer(consumer);
-        consumerVerticle.setCloser(AsyncCloseable.wrapAutoCloseable(metricsCloser));
-      })
+          // Set all the built objects in the consumer verticle
+          consumerVerticle.setRecordDispatcher(recordDispatcher);
+          consumerVerticle.setConsumer(consumer);
+          consumerVerticle.setCloser(AsyncCloseable.wrapAutoCloseable(metricsCloser));
+        })
         .mapEmpty();
 
     return getConsumerVerticle(deliveryOrder, initializer, new HashSet<>(resource.getTopicsList()));
@@ -217,16 +217,16 @@ public class ConsumerVerticleFactoryImpl implements ConsumerVerticleFactory {
     // Dialected filters should override the attributes filter
     if (egress.getDialectedFilterCount() > 0) {
       return getFilter(egress.getDialectedFilterList());
-    } else if (egress.hasFilter()){
+    } else if (egress.hasFilter()) {
       return new AttributesFilter(egress.getFilter().getAttributesMap());
     }
     return Filter.noop();
   }
 
   private Filter getFilter(DataPlaneContract.DialectedFilter filter) {
-    return switch (filter.getFilterCase()){
+    return switch (filter.getFilterCase()) {
       case EXACT -> new AttributesFilter(filter.getExact().getAttributesMap());
-      case PREFIX -> new PrefixFilter(filter.getPrefix().getAttribute(),filter.getPrefix().getPrefix());
+      case PREFIX -> new PrefixFilter(filter.getPrefix().getAttribute(), filter.getPrefix().getPrefix());
       case SUFFIX -> new SuffixFilter(filter.getSuffix().getAttribute(), filter.getSuffix().getSuffix());
       case NOT -> new NotFilter(getFilter(filter.getNot().getFilter()));
       case ANY -> new AnyFilter(filter.getAny().getFiltersList().stream().
