+++ data-plane/dispatcher/src/test/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/subscriptionsapi/SuffixFilterTest.java
@@ -13,24 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi;
 
-import dev.knative.eventing.kafka.broker.dispatcher.Filter;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
-import java.net.URI;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
-import java.util.Collections;
-import java.util.Set;
-import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.net.URI;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
+import java.util.stream.Stream;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class AllFilterTest {
+public class SuffixFilterTest {
 
   final static CloudEvent event = CloudEventBuilder.v1()
     .withId("123-42")
@@ -48,18 +46,18 @@ public class AllFilterTest {
 
   @ParameterizedTest
   @MethodSource(value = {"testCases"})
-  public void match(CloudEvent event, Filter filter, boolean shouldMatch) {
+  public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
+    var filter = new SuffixFilter(key, value);
     assertThat(filter.test(event))
       .isEqualTo(shouldMatch);
   }
 
   static Stream<Arguments> testCases() {
     return Stream.of(
-      Arguments.of(event, new AllFilter(Set.of(new ExactFilter("id", "123-42"))), true),
-      Arguments.of(event, new AllFilter(Set.of(new ExactFilter("id", "123-42"), new ExactFilter("source", "/api/some-source"))), true),
-      Arguments.of(event, new AllFilter(Set.of(new ExactFilter("id", "123"), new ExactFilter("source", "/api/some-source"))), false),
-      Arguments.of(event, new AllFilter(Set.of(new ExactFilter("id", "123-42"), new ExactFilter("source", "/api/something-else"))), false),
-      Arguments.of(event, new AllFilter(Collections.emptySet()), true)
+      Arguments.of(event, "id", "42", true),
+      Arguments.of(event, "id", "43", false),
+      Arguments.of(event, "source", "source", true),
+      Arguments.of(event, "source", "sink", false)
     );
   }
